Projects/Batch-2022-2026/160922729008-Predictive Modeling of Drug Safety Using ADR Data/PREDICTIVE MODELLING OF DRUG SAFETY USING ADVERSE DRUG REACTION DATA/backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
joblib
pandas
mysql-connector-python
python-dotenv
hypothesis
pytest
httpx
